Add files via upload
The “E-Learning Website” has been developed to override the
problems prevailing in the practicing manual system. This website
is supported to eliminate and in some cases reduce the hardships
faced by this existing system. Moreover this system is designed for
the particular need of the administrator to carry out operations in
a smooth and effective manner and the user to make selections
smoothly.